unit user, improve

+++ app/Http/Controllers/Admin/UnitController.php
@@ -4,19 +4,24 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Unit;
+use App\Models\UnitUser;
 use Illuminate\Http\Request;
 
 class UnitController extends Controller
 {
     public function index()
     {
+        $data['sidebar'] = 'unit';
+        $data['sidebar_sub'] = 'units';
          $data['units'] = Unit::get();
          return view('admin.units.index', $data);
     }
 
     public function create()
     {
-         return view('admin.units.create');
+        $data['sidebar'] = 'unit';
+        $data['sidebar_sub'] = 'unit_create';
+         return view('admin.units.create', $data);
     }
 
     public function store(Request $request)
@@ -24,10 +29,10 @@ class UnitController extends Controller
         $request['status'] = 'active';
 
         if(Unit::create($request->all())) {
-            return redirect()->route('admin.units')->with('success', 'Unit created successfully!');
+            return redirect()->route('admin.units')->with('success', actionMessage());
         }
 
-        return redirect()->back()->with('error', 'Something went wrong! Try Again');
+        return redirect()->back()->with('error', actionMessage('error'));
     }
 
     public function edit($id)
@@ -36,11 +41,13 @@ class UnitController extends Controller
 
         if (!empty($unit)) {
 
+            $data['sidebar'] = 'unit';
+            $data['sidebar_sub'] = 'units';
             $data['unit'] = $unit;
             return view('admin.units.edit', $data);
         }
 
-        return redirect()->back()->with('error', 'Unit not found!');
+        return redirect()->back()->with('error', actionMessage('notFound'));
     }
 
     public function update(Request $request, $id)
@@ -52,10 +59,10 @@ class UnitController extends Controller
             $unit->params = $request->params;
             $unit->save();
 
-            return redirect()->route('admin.units')->with('success', 'Unit updated successfully!');
+            return redirect()->route('admin.units')->with('success', actionMessage('update'));
         }
 
-        return redirect()->back()->with('error', 'Something went wrong!');
+        return redirect()->back()->with('error', actionMessage('error'));
 
     }
 
@@ -64,15 +71,27 @@ class UnitController extends Controller
         $unit = Unit::where('id', $request->id)->first();
 
         if(!empty($unit)) {
+
+            $uniUser = UnitUser::where('unit_id', $unit->id);
+
+            if($uniUser->exists()) {
+                return response()->json([
+                    'status' => 0,
+                    'message' => actionMessage('failCustom'),
+                ]);
+            }
+
             $unit->delete();
 
             return response()->json([
-                'status' => 1
+                'status' => 1,
+                'message' => actionMessage('delete'),
             ]);
         }
 
         return response()->json([
-                'status' => 0
+                'status' => 0,
+                'message' => actionMessage('error'),
             ]);
     }
 }
